+++ tiledb/sm/query/query.cc
@@ -199,6 +199,77 @@ Status Query::get_range_var(
   return Status::Ok();
 }
 
+Status Query::add_range_by_name(
+    const std::string& dim_name,
+    const void* start,
+    const void* end,
+    const void* stride) {
+  unsigned dim_idx;
+  RETURN_NOT_OK(array_->array_schema()->domain()->get_dimension_index(
+      dim_name, &dim_idx));
+
+  return add_range(dim_idx, start, end, stride);
+}
+
+Status Query::add_range_var_by_name(
+    const std::string& dim_name,
+    const void* start,
+    uint64_t start_size,
+    const void* end,
+    uint64_t end_size) {
+  unsigned dim_idx;
+  RETURN_NOT_OK(array_->array_schema()->domain()->get_dimension_index(
+      dim_name, &dim_idx));
+
+  return add_range_var(dim_idx, start, start_size, end, end_size);
+}
+
+Status Query::get_range_num_from_name(
+    const std::string& dim_name, uint64_t* range_num) const {
+  unsigned dim_idx;
+  RETURN_NOT_OK(array_->array_schema()->domain()->get_dimension_index(
+      dim_name, &dim_idx));
+
+  return get_range_num(dim_idx, range_num);
+}
+
+Status Query::get_range_from_name(
+    const std::string& dim_name,
+    uint64_t range_idx,
+    const void** start,
+    const void** end,
+    const void** stride) const {
+  unsigned dim_idx;
+  RETURN_NOT_OK(array_->array_schema()->domain()->get_dimension_index(
+      dim_name, &dim_idx));
+
+  return get_range(dim_idx, range_idx, start, end, stride);
+}
+
+Status Query::get_range_var_size_from_name(
+    const std::string& dim_name,
+    uint64_t range_idx,
+    uint64_t* start_size,
+    uint64_t* end_size) const {
+  unsigned dim_idx;
+  RETURN_NOT_OK(array_->array_schema()->domain()->get_dimension_index(
+      dim_name, &dim_idx));
+
+  return get_range_var_size(dim_idx, range_idx, start_size, end_size);
+}
+
+Status Query::get_range_var_from_name(
+    const std::string& dim_name,
+    uint64_t range_idx,
+    void* start,
+    void* end) const {
+  unsigned dim_idx;
+  RETURN_NOT_OK(array_->array_schema()->domain()->get_dimension_index(
+      dim_name, &dim_idx));
+
+  return get_range_var(dim_idx, range_idx, start, end);
+}
+
 Status Query::get_est_result_size(const char* name, uint64_t* size) {
   if (type_ == QueryType::WRITE)
     return LOG_STATUS(Status::QueryError(